Made the poll a child of the PollList component.
